added code for parsing the messages between the client and server using.
updated the handleMessage and broadcastGameState methods to now use the type and data fields to align with the GameMessage interface on the client side websocket file.
The code now uses data.getString() and data.getJSONObject() to extract the specific data from the data JSONObject.
the create and join message handling was edited to match the json object structure in the websocket file on the client side.