Update eslint configuration
This addresses a failing build due to `ng lint` claiming that the use of == is incorrect here. Locally, this still shows the same as an error, but I want to see if it behaves differently on github actions. It seems like this rule may need to be off rather than "warn", but I want to be sure.