W.I.P. commit for the Rejoin function
Within this commit, I've created part of the framework for the rejoin functionality. More specifically, I've added a few calls within the server to trigger the delivery of a message to the client who left. Within the client's session, the session then sends back an updated boolean array titled ConnectedClients, which keeps track of the clients connected to the game.

There's some issues with getting the connectedPlayers boolean to attribute the right person who left; that's what I'm working on currently.