Update E2E testing to use cy.request
This Cypress testing code includes direct manipulation of the DOM of the game page in the Cypress test using a PUT request through cy.request. Then, it checks that the page updated appropriately.

The addition of tasks in the cypress configuration seems unneeded given that the web socket messages are not used so that specific things are updated. Instead, the presence of a web socket message just prompts the game page to refresh, which will mean it has the most up to date information. We left those in for now so that it is possible to see how to use custom tasks, but the cases tested so far don't really need it.

We added a new game object that contains more information to the seed file for games. We also then *use* that new game object.

If it's simpler for adding more tests to make new Cypress test files that don't use page objects, the team could consider that approach. You might also want to make other game objects with particular properties. There are many things that have not been checked in this E2E test right now, but this commit definitely includes some useful demonstrations about how to use cy.request.

Co-authored-by: James Swearingen <swear051@morris.umn.edu>