Fixed Websocket Delay
Within this commit, I fixed a delay caused when running the websocket code by removing the implementation of the CLIENT_ALIVE_STATUS hashmap. As the functionality from this can get done by checking for client.session.isOpen() instead in broadcastUpdate(), this variable seems unneeded. Furthermore, as the purpose of a heartbeat lies in keeping a websocket connection open, the ping itself checking for who's connected ends up putting a significant amount of strain on the speed of the websockets.